Inline Filter refactoring.
Re-write the inline filter implementation using Cheetah + Jquery, in place
of generating an enormous amount of javascript code. This drastically reduce
the size of the events listing HTML page, and will allow for much easier
modification of the inline-filters. Additionally, we now only propose filter
operator relevant to the selected path.