One click filter on classification.
Improve Prewikka behavior, by making a click on classification add a filter
on the selected classification (previously, it would have unfolded aggregated
alerts for the selected entry - this can now be done clicking the alert count).

Depending on the context; the user might be presented with a choise,  allowing
him to see the clicked alert details, or to set a filter on the classification.