Implement simplified inline filter.
Inline filter now present a single input field (with no path and operator
selection). Using this field, the user can filter on what is seen in the
associated column. For example, in the classification column, the filter
will trigger a search on classification.text, classification.reference.name
and classification.reference.origin.

There is also an [expert] button allowing the user to specify both the
path and the operator.