Inline filters improvements: reset switch, and distinction between applied/saved filters.
Implement a reset button in each inline filter column, that allow to switch
between different version of the filter: last saved filters, default filters,
or current filters.

Make the "[filtered]" information black when the filter has been saved as part
of the view, this make a difference between saved, and only applied filters.