Sources and Targets expansion link fixes
The link used to expand a list of sources/target was always broken. The link
now point to the detailed view for CorrelationAlert, or the detailed event
for alert.

Additionally, this drop _fetchInfoFromLinkedMessage(), which was used to gather
information for very old Correlator version whose events version didn't set
sources and targets information.