Fixes possible Inline-filter null reset button
When the inline filter reset button is clicked, do not try to detect
what the current state is, since the users might have modified the
filters, leading to an unknown state.

There is no need for this detection since the next state was already
retrieved on the previous call, so we re-use the value.

This prevent a rare 'Reset to null' button which could happen when
modifying the inline-filter (in a specific way) before clicking the
Reset button.