[#rel1094] Add support for reverse proxies
Add the "reverse_path" configuration option in prewikka.conf which can
be used to override the base path to prewikka.

This is useful when the application is being served through a reverse
proxy with a different prefix than on the backend server.

This setting is then used in redirections, path lookups, etc.

Change-Id: Ia8ba72b441453ec8f249a301f04377cf34c63a72