[#rel1701] Standardize user/request access
View used to hold reference to the user and request object, but theses attributes
really did not belong here. This also lead to unclear code where the view was passed
as an argument to various hook for the hook to access the underlying attribute.

The standard is now to access theses attribute globally, through the thread local
env.request object.

Change-Id: Ib86b5297f5546623d9f3c28ff8c7f24155264189