[#3358] YAML menu relative to main config file
Make it so that the "menu_order" configuration directive becomes
relative to the main configuration file.

Previously, the "menu_order" directive was always interpreted as
relative to the configuration folder set during the installation
(usually "/etc/prewikka/"), even when another file was set to be
used as the main configuration file using the PREWIKKA_CONFIG
(WSGI) environment variable.

Change-Id: Ie31f382cb4bef88536ade078abedcf800a27e71b